Add ILogger logging
For users and to help pin down test failures. AI is guessing and needs
more knowledge of internal flows.

This is mostly AI written though it took a few attempts to make it do it
right and not break everything. For example it started with console
write which is no good for a lib.